tools/project_gen/requirements.txt
celery==5.0.5
GitPython==3.1.13
numpy==1.20.1
pandas==1.2.2
PyYAML==5.4.1